+++ tests/functional/pytorch/016_aitune_torch_tune_stable_diffusion_tensorrt_with_onnx_dynamo_test.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 # /// script
-# dependencies = ["transformers<5", "diffusers"]
+# dependencies = ["transformers", "diffusers"]
 #
 # # Optional, default "always", determines how often test is generated, always, nightly, weekly, monthly
 # scope = "always"